+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Features
 
+* **bable:** add bable-polyfill and babel-preset-stage-2 ([69514cb](https://github.com/pfan123/front-end-navigator/commit/69514cb))
 * **commitValidate:** add husky and validate-commit-msg use as a githooks to validate the commit message ([5dbbf3e](https://github.com/pfan123/front-end-navigator/commit/5dbbf3e))
 * **modifyreadme:** add webpack in readme ([ea8b9b8](https://github.com/pfan123/front-end-navigator/commit/ea8b9b8))
 * **test:** test new Vue and components different ([51ae20d](https://github.com/pfan123/front-end-navigator/commit/51ae20d))